+++ app/src/main/AndroidManifest.xml
@@ -20,12 +20,12 @@
             android:label="@string/title_activity_my"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.auth.SignUpActivity"
+            android:name=".presentation.screen.signUp.SignUpActivity"
             android:exported="false"
             android:label="@string/title_activity_sign_up"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".ui.auth.SignInActivity"
+            android:name=".presentation.screen.signIn.SignInActivity"
             android:exported="true"
             android:label="@string/title_activity_sign_in"
             android:theme="@style/Theme.ATSOPTANDROID">
@@ -35,7 +35,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ATSOPTANDROID">
