+++ app/src/main/java/org/sopt/at/Color.kt
@@ -1,0 +1,10 @@
+package org.sopt.at
+
+
+import androidx.compose.ui.graphics.Color
+
+val TvingBlack = Color(0xFF000000)
+val TvingWhite = Color(0xFFFFFFFF)
+val TvingRed = Color(0xFFFF143C)
+
+
