+++ app/src/main/java/org/sopt/at/data/dto/response/SignUpResponseDto.kt
@@ -2,6 +2,7 @@ package org.sopt.at.data.dto.response
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
+import org.sopt.at.domain.model.UserEntity
 
 @Serializable
 data class SignUpResponseDto(
@@ -9,4 +10,11 @@ data class SignUpResponseDto(
     val userId: Int,
     @SerialName("nickname")
     val nickname: String
-)
+) {
+    fun SignUpResponseDto.toEntity(): UserEntity {
+        return UserEntity(
+            userId = userId,
+            nickname = nickname
+        )
+    }
+}
