--- app/src/main/java/org/sopt/at/data/network/RetrofitInstance.kt
@@ -1,27 +1,0 @@
-package org.sopt.at.data.network
-
-import org.sopt.at.BuildConfig
-import org.sopt.at.data.service.AuthService
-import org.sopt.at.data.service.UserService
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-    private val BASE_URL = BuildConfig.BASE_URL
-
-    val authService: AuthService by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(AuthService::class.java)
-    }
-    val userService: UserService by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(UserService::class.java)
-
-    }
-}
