+++ app/src/main/java/org/sopt/at/domain/model/UserEntity.kt
@@ -1,0 +1,6 @@
+package org.sopt.at.domain.model
+
+data class UserEntity(
+    val userId: Int,
+    val nickname: String? = null
+)
