+++ app/src/main/java/org/sopt/at/presentation/screen/history/HistoryScreen.kt
@@ -1,0 +1,22 @@
+package org.sopt.at.presentation.screen.history
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun HistoryScreen() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colorScheme.background),
+        contentAlignment = Alignment.Center
+    ) {
+        Text("History Screen", color = MaterialTheme.colorScheme.onBackground)
+    }
+}
