+++ app/src/main/java/org/sopt/at/presentation/screen/my/MyScreen.kt
@@ -1,0 +1,245 @@
+package org.sopt.at.presentation.screen.my
+
+import android.content.Intent
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ChevronLeft
+import androidx.compose.material.icons.filled.ChevronRight
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavController
+import org.sopt.at.R
+import org.sopt.at.presentation.screen.signIn.SignInActivity
+
+
+@Composable
+fun MyScreen(navController: NavController, viewModel: MyViewModel = viewModel()) {
+    val context = LocalContext.current
+    LaunchedEffect(Unit) {
+        viewModel.fetchNickname(context)
+    }
+    Profile(nickname = viewModel.nickname, navController = navController)
+}
+
+
+@Composable
+fun Profile(nickname: String, navController: NavController) {
+    val context = LocalContext.current
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colorScheme.background)
+
+    ) {
+        Icon(
+            modifier = Modifier
+                .size(35.dp)
+                .clickable{
+                    navController.navigate("home")
+                },
+            imageVector = Icons.Default.ChevronLeft,
+            contentDescription = null,
+            tint = MaterialTheme.colorScheme.onBackground,
+        )
+
+        Spacer(modifier = Modifier.height(20.dp))
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .background(MaterialTheme.colorScheme.background)
+                .padding(horizontal = 25.dp, vertical = 16.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.Start
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.ic_tving),
+                    contentDescription = null,
+                    modifier = Modifier
+                        .size(60.dp)
+                        .background(Color.Yellow)
+                )
+                Spacer(modifier = Modifier.width(18.dp))
+                Text(
+                    text = nickname,
+                    color = MaterialTheme.colorScheme.onBackground,
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+            Button(
+                onClick = {
+                    navController.navigate("changeNickname")
+                },
+                modifier = Modifier
+                    .border(width = 1.dp, color = Color.Gray, shape = RoundedCornerShape(4.dp))
+                    .height(35.dp)
+                    .width(112.dp),
+                colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.surface)
+            ) {
+                Text(
+                    "프로필 전환",
+                    color = MaterialTheme.colorScheme.outline,
+                    style = MaterialTheme.typography.labelSmall
+                )
+            }
+        }
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 20.dp, horizontal = 16.dp),
+            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.background),
+            border = BorderStroke(1.dp, MaterialTheme.colorScheme.outline),
+            shape = RoundedCornerShape(5.dp)
+        ) {
+            Column(
+                modifier = Modifier
+                    .padding(vertical = 28.dp, horizontal = 20.dp)
+            ){
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ){
+                    Text(
+                        text="나의 이용권",
+                        color = MaterialTheme.colorScheme.onBackground,
+                        fontSize = 14.sp
+                    )
+                    Text(
+                        text = "사용중인 이용권이 없습니다",
+                        color = MaterialTheme.colorScheme.onBackground,
+                        style = MaterialTheme.typography.titleMedium
+                    )
+                }
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 16.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    Text(
+                        text= "티빙캐시",
+                        color = MaterialTheme.colorScheme.onSurface,
+                        fontSize = 14.sp)
+                    Text(
+                        text="0",
+                        color = MaterialTheme.colorScheme.onSurface,
+
+                        )
+                }
+
+            }
+        }
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 12.dp, vertical = 6.dp),
+            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.background),
+            shape = RoundedCornerShape(5.dp),
+            border = BorderStroke(1.dp, MaterialTheme.colorScheme.outline),
+        ) {
+            Column(
+                modifier = Modifier
+                    .padding(vertical = 5.dp, horizontal = 20.dp)
+            ) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Column() {
+                        Row(
+                            horizontalArrangement = Arrangement.Start,
+                            verticalAlignment = Alignment.CenterVertically
+                        ) {
+                            Text(
+                                text = "이용권을 구매하고  ",
+                                style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.onSurface)
+
+                            Image(
+                                painter = painterResource(id = R.drawable.ic_tvn),
+                                contentDescription = null,
+                                modifier = Modifier
+                                    .size(37.dp)
+                                    .background(Color.Black)
+                            )
+                            Spacer(modifier = Modifier.width(5.dp))
+                            Image(
+                                painter = painterResource(id = R.drawable.ic_jtbc),
+                                contentDescription = null,
+                                modifier = Modifier
+                                    .size(52.dp)
+                                    .background(Color(0xFF1A1A1A))
+                            )
+                            Text(
+                                text=" 등 인기 시리즈와",
+                                color = MaterialTheme.colorScheme.onSurface,
+                                style = MaterialTheme.typography.labelSmall
+                            )
+                        }
+                        Text(
+                            text = "다양한 영화 콘텐츠를 자유롭게 시청하세요!",
+                            color = MaterialTheme.colorScheme.onSurface,
+                            style = MaterialTheme.typography.labelSmall
+                        )
+                    }
+                    Icon(
+                        imageVector = Icons.Default.ChevronRight,
+                        contentDescription = null,
+                        tint = MaterialTheme.colorScheme.onSurface
+                    )
+                }
+
+            }
+
+        }
+        Button(
+            onClick ={
+                val intent = Intent(context, SignInActivity::class.java)
+                context.startActivity(intent)
+            },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 24.dp, vertical = 32.dp)
+                .border(width = 1.dp, color = Color.Gray, shape = RoundedCornerShape(4.dp))
+                .height(48.dp),
+            colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.background)
+        ) {
+            Text("로그아웃", color = MaterialTheme.colorScheme.outline)
+        }
+    }
+}
