+++ app/src/main/java/org/sopt/at/presentation/screen/search/SearchScreen.kt
@@ -1,0 +1,57 @@
+package org.sopt.at.presentation.screen.search
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.compose.foundation.lazy.items
+
+@Composable
+fun SearchScreen(viewModel: SearchViewModel = viewModel()) {
+    var input by remember { mutableStateOf("")}
+
+    Column(
+        modifier = Modifier
+            .padding(16.dp)
+
+    ){
+        OutlinedTextField(
+            value = input,
+            onValueChange = {
+                input = it
+                viewModel.searchNickname(it)
+            },
+            label = {Text("닉네임 검색")},
+            modifier = Modifier.fillMaxWidth()
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+
+        LazyColumn(
+            modifier = Modifier.fillMaxHeight()
+        ) {
+            items(viewModel.nicknameList) { name ->
+                Text(
+                    text = name,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 8.dp),
+                    style = MaterialTheme.typography.bodyLarge
+                )
+            }
+        }
+    }
+}
