+++ app/src/main/java/org/sopt/at/presentation/screen/shorts/ShortsScreen.kt
@@ -1,0 +1,22 @@
+package org.sopt.at.presentation.screen.shorts
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun ShortsScreen() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.Black),
+        contentAlignment = Alignment.Center
+    ) {
+        Text("Shorts Screen", color = Color.White)
+    }
+}
