--- app/src/main/java/org/sopt/at/screen/ui/theme/Color.kt
@@ -1,10 +1,0 @@
-package org.sopt.at.screen.ui.theme
-
-
-import androidx.compose.ui.graphics.Color
-
-val TvingBlack = Color(0xFF000000)
-val TvingWhite = Color(0xFFFFFFFF)
-val TvingRed = Color(0xFFFF143C)
-
-
