--- app/src/main/java/org/sopt/at/ui/home/HomeViewModel.kt
@@ -1,50 +1,0 @@
-package org.sopt.at.ui.home
-
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.lifecycle.ViewModel
-import org.sopt.at.R
-import org.sopt.at.model.Top20
-
-class HomeViewModel : ViewModel() {
-    val genres = listOf("DRAMA", "VARIETY", "MOVIE", "SPORTS", "ANIMATION")
-
-    var selectedGenre by mutableStateOf(genres.first())
-
-    val currentContents by derivedStateOf {
-        genreContentMap[selectedGenre] ?: emptyList()
-    }
-
-    fun onGenreSelected(genre: String) {
-        selectedGenre = genre
-    }
-
-    val genreContentMap = mapOf(
-        "DRAMA" to listOf(R.drawable.banner1, R.drawable.banner1),
-        "VARIETY" to listOf(R.drawable.variety, R.drawable.variety),
-        "MOVIE" to listOf(R.drawable.movie, R.drawable.movie),
-        "SPORTS" to listOf(R.drawable.sports1, R.drawable.sports1),
-        "ANIMATION" to listOf(R.drawable.animation, R.drawable.animation)
-    )
-
-    val topList = listOf(
-        Top20(1, R.drawable.top1),
-        Top20(2, R.drawable.top2),
-        Top20(3, R.drawable.top2),
-        Top20(4, R.drawable.top1),
-        Top20(5, R.drawable.top1),
-        Top20(6, R.drawable.top1),
-        Top20(7, R.drawable.top1)
-    )
-
-    val contentsList = listOf(
-        R.drawable.now1,
-        R.drawable.now2,
-        R.drawable.now3,
-        R.drawable.now1,
-        R.drawable.now1,
-        R.drawable.now1
-    )
-}
