--- app/src/main/java/org/sopt/at/ui/main/MainActivity.kt
@@ -1,26 +1,0 @@
-package org.sopt.at.ui.main
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import org.sopt.at.screen.ui.theme.ATSOPTANDROIDTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        val nickname = intent.getStringExtra("nickname") ?: "티빙"
-        setContent {
-            ATSOPTANDROIDTheme {
-                MainScreen(nickname = nickname)
-
-            }
-        }
-    }
-}
-
